Split pipeline proposals per file
